LayoutTests/platform/chromium/http/tests/inspector/console-xhr-logging-expected.txt

CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
Tests that XMLHttpRequest Logging works when Enabled and doesn't show logs when Disabled.

sending a GET request to resources/xhr-exists.html console-xhr-logging.html:10
XHR finished loading: "http://127.0.0.1:8000/inspector/resources/xhr-exists.html". network-test.js:44
sending a GET request to resources/xhr-does-not-exist.html console-xhr-logging.html:10
GET http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html 404 (Not Found) network-test.js:44
XHR finished loading: "http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html". network-test.js:44
sending a POST request to resources/post-target.cgi console-xhr-logging.html:10
XHR finished loading: "http://127.0.0.1:8000/inspector/resources/post-target.cgi". network-test.js:44
sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html console-xhr-logging.html:10
XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
sending a GET request to resources/xhr-exists.html console-xhr-logging.html:10
sending a GET request to resources/xhr-does-not-exist.html console-xhr-logging.html:10
GET http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html 404 (Not Found) network-test.js:44
sending a POST request to resources/post-target.cgi console-xhr-logging.html:10
sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html console-xhr-logging.html:10
XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
